A big fat refactor
