+++ src/main/java/com/juiceybeans/scorched_claims/core/impl/ChunkPowerImpl.java
@@ -1,6 +1,13 @@
-package com.juiceybeans.scorched_claims.api;
+package com.juiceybeans.scorched_claims.core.impl;
+
+import com.juiceybeans.scorched_claims.api.IChunkPower;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.CapabilityManager;
+import net.minecraftforge.common.capabilities.CapabilityToken;
 
 public class ChunkPowerImpl implements IChunkPower {
+    public static final Capability<IChunkPower> INSTANCE = CapabilityManager.get(new CapabilityToken<>() {});
+
     private int power = 1000;
 
     @Override
