+++ src/main/resources/scorched_claims.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "com.juiceybeans.scorched_claims.core.mixin",
+  "package": "com.juiceybeans.scorched_claims.mixin",
   "compatibilityLevel": "JAVA_17",
   "refmap": "mixins.scorched_claims.refmap.json",
   "mixins": [
